+++ cmd/admin_tool/cmd.go
@@ -1,0 +1,8 @@
+package admin_tool
+
+import "github.com/spf13/cobra"
+
+var Cmd = cobra.Command{
+	Use:   "admin-tool",
+	Short: "Debug utility for administering hive components",
+}
