+++ cmd/providers/authgw.go
@@ -1,83 +1,44 @@
-package main
+package providers
 
 import (
 	"context"
 	"encoding/hex"
+	"time"
 
 	"github.com/go-redis/redis/v8"
+	lru "github.com/hashicorp/golang-lru"
 	"github.com/jmoiron/sqlx"
 	"github.com/spf13/viper"
 	"go.od2.network/hive/pkg/auth"
 	"go.od2.network/hive/pkg/authgw"
-	"go.od2.network/hive/pkg/njobs"
+	"go.od2.network/hive/pkg/cachegc"
 	"go.od2.network/hive/pkg/token"
 	"go.uber.org/fx"
 	"go.uber.org/zap"
 )
 
-var providers = []interface{}{
-	newRedis,
-	newRedisNJobs,
-	newMySQL,
-	newAuthgwBackend,
-	newAuthgwSigner,
-	newWorkerAuthInterceptor,
+const (
+	ConfAuthgwSecret         = "authgw.secret"
+	ConfAuthgwCacheSize      = "authgw.cache.size"
+	ConfAuthgwCacheTTL       = "authgw.cache.ttl"
+	ConfAuthgwCacheStreamKey = "authgw.cache.stream_key"
+	ConfAuthgwCacheBacklog   = "authgw.cache.backlog"
+)
+
+func init() {
+	viper.SetDefault(ConfAuthgwSecret, "")
+	viper.SetDefault(ConfAuthgwCacheSize, 1024)
+	viper.SetDefault(ConfAuthgwCacheTTL, time.Hour)
+	viper.SetDefault(ConfAuthgwCacheStreamKey, "token-invalidations")
+	viper.SetDefault(ConfAuthgwCacheBacklog, 64)
 }
 
-func newRedis(ctx context.Context, lc fx.Lifecycle) (*redis.Client, error) {
-	rd := redisClientFromEnv()
-	if err := rd.Ping(ctx).Err(); err != nil {
-		return nil, err
-	}
-	lc.Append(fx.Hook{
-		OnStop: func(ctx context.Context) error {
-			log.Info("Closing Redis client")
-			err := rd.Close()
-			if err != nil {
-				log.Error("Failed to close Redis client", zap.Error(err))
-			}
-			return err
-		},
-	})
-	return rd, nil
-}
-
-func newRedisNJobs(ctx context.Context, rd *redis.Client) (*njobs.RedisClient, error) {
-	scripts, err := njobs.LoadScripts(ctx, rd)
-	if err != nil {
-		log.Fatal("Failed to load njobs scripts", zap.Error(err))
-	}
-	// Connect to Redis njobs.
-	topic, partition := kafkaPartitionFromEnv()
-	return &njobs.RedisClient{
-		Redis:         rd,
-		PartitionKeys: njobs.NewPartitionKeys(topic, partition),
-		Scripts:       scripts,
-		Options:       njobsOptionsFromEnv(),
-	}, nil
-}
-
-func newMySQL(ctx context.Context, lc fx.Lifecycle) (*sqlx.DB, error) {
-	db, err := openDB()
-	if err != nil {
-		return nil, err
-	}
-	if err := db.PingContext(ctx); err != nil {
-		log.Fatal("Failed to ping DB", zap.Error(err))
-	}
-	lc.Append(fx.Hook{
-		OnStop: func(ctx context.Context) error {
-			return db.Close()
-		},
-	})
-	return db, nil
-}
-
-func newAuthgwBackend(
+func NewAuthgwBackend(
 	lc fx.Lifecycle,
 	shutdown fx.Shutdowner,
 	db *sqlx.DB,
 	rd *redis.Client,
+	log *zap.Logger,
 ) (authgw.Backend, error) {
 	// Build auth gateway.
 	backend := authgw.Database{DB: db.DB}
@@ -118,7 +79,7 @@ func newAuthgwBackend(
 	return cachedBackend, nil
 }
 
-func newAuthgwSigner() token.Signer {
+func NewAuthgwSigner(log *zap.Logger) token.Signer {
 	authSecret := new([32]byte)
 	authSecretStr := viper.GetString(ConfAuthgwSecret)
 	if len(authSecretStr) != 64 {
@@ -130,10 +91,27 @@ func newAuthgwSigner() token.Signer {
 	return token.NewSimpleSigner(authSecret)
 }
 
-func newWorkerAuthInterceptor(backend authgw.Backend, signer token.Signer) *auth.WorkerAuthInterceptor {
+func NewWorkerAuthInterceptor(
+	log *zap.Logger,
+	backend authgw.Backend,
+	signer token.Signer,
+) *auth.WorkerAuthInterceptor {
 	return &auth.WorkerAuthInterceptor{
 		Backend: backend,
 		Signer:  signer,
 		Log:     log.Named("auth"),
 	}
+}
+
+type AuthgwCache struct {
+	*cachegc.Cache
+}
+
+func NewAuthgwCache() AuthgwCache {
+	baseCache, err := lru.New(viper.GetInt(ConfAuthgwCacheSize))
+	if err != nil {
+		panic("failed to create cache: " + err.Error())
+	}
+	cache := cachegc.NewCache(baseCache, viper.GetDuration(ConfAuthgwCacheTTL))
+	return AuthgwCache{cache}
 }
